+++ src/components/themed/SceneHeaderUi4.tsx
@@ -42,9 +42,9 @@ const getStyles = cacheStyles((theme: Theme) => ({
     paddingBottom: theme.rem(DEFAULT_MARGIN_REM)
   },
   titleContainer: {
+    alignItems: 'center',
     flexDirection: 'row',
     justifyContent: 'space-between',
-    alignItems: 'center',
     marginHorizontal: theme.rem(DEFAULT_MARGIN_REM),
     marginBottom: theme.rem(DEFAULT_MARGIN_REM)
   },